chore: a new react-ts app was created with initial configuration
This project is ready to work with linters, prettier, jest and testing library; enforcing
develoment and quality code standards with the help of husky and conventional commits. Feel free to
go and read the README.md file available for further details. IMPORTANT: Upcoming changes will be
introduced as we go shaping progressively the boilerplate and adding more tools to work with.